Replace token logic with session logic, update DB, and refactor code

+++ VibeHub/API/Controllers/GuestController.cs
@@ -1,7 +1,9 @@
 using BLL.Abstractions.Services;
+using Microsoft.AspNetCore.Authentication.Cookies;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace API.Controllers;
 
@@ -10,31 +12,55 @@ namespace API.Controllers;
 public class GuestController : ControllerBase
 {
     private readonly IGuestService _guestService;
-    private readonly ITokenService _tokenService;
+    private readonly ISessionService _sessionService;
 
-    public GuestController(IGuestService guestService, ITokenService tokenService)
+    public GuestController(IGuestService guestService, ISessionService sessionService)
     {
         _guestService = guestService;
-        _tokenService = tokenService;
+        _sessionService = sessionService;
     }
     
     [HttpPost]
     public async Task<IActionResult> Create()
     {
         var guest = await _guestService.Create();
-        var token = _tokenService.GenerateToken(null, guest);
+        var sessionId = _sessionService.CreateSession(null, guest);
 
-        return Ok(new { Token = token });
+        HttpContext.Session.SetString("SessionId", sessionId);
+
+        var claims = new List<Claim>
+        {
+            new Claim(ClaimTypes.NameIdentifier, sessionId)
+        };
+
+        var claimsIdentity = new ClaimsIdentity(claims, CookieAuthenticationDefaults.AuthenticationScheme);
+        var authProperties = new AuthenticationProperties
+        {
+            IsPersistent = true
+        };
+
+        await HttpContext.SignInAsync(CookieAuthenticationDefaults.AuthenticationScheme, new ClaimsPrincipal(claimsIdentity), authProperties);
+
+        return Ok(new { SessionId = sessionId });
     }
 
     [Authorize]
     [HttpDelete]
     public async Task<IActionResult> Delete()
     {
-        var guestId = _tokenService.GetIdFromToken();
+        var sessionId = HttpContext.Session.GetString("SessionId");
+
+        if (string.IsNullOrEmpty(sessionId))
+        {
+            return Unauthorized();
+        }
+
+        var guestId = _sessionService.GetGuestIdFromSession(sessionId);
         await _guestService.Delete(guestId);
-        
+
+        HttpContext.Session.Remove("SessionId");
+        await HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
+
         return NoContent();
     }
-
 }
